Moved JSONString out of split package
