+++ src/main/java/org/skyscreamer/jsonassert/JSONCompare.java
@@ -17,7 +17,6 @@ package org.skyscreamer.jsonassert;
 import org.json.JSONArray;
 import org.json.JSONException;
 import org.json.JSONObject;
-import org.json.JSONString;
 import org.skyscreamer.jsonassert.comparator.DefaultComparator;
 import org.skyscreamer.jsonassert.comparator.JSONComparator;
 
@@ -95,7 +94,7 @@ public final class JSONCompare {
 
     /**
      * Compares {@link JSONString} provided to the expected {@code JSONString}, checking that the
-     * {@link org.json.JSONString#toJSONString()} are equal.
+     * {@link JSONString#toJSONString()} are equal.
      *
      * @param expected Expected {@code JSONstring}
      * @param actual   {@code JSONstring} to compare
