+++ src/main/java/org/skyscreamer/jsonassert/JSONParser.java
@@ -17,7 +17,6 @@ package org.skyscreamer.jsonassert;
 import org.json.JSONArray;
 import org.json.JSONException;
 import org.json.JSONObject;
-import org.json.JSONString;
 
 /**
  * Simple JSON parsing utility.
